+++ src/components/Secret.js
@@ -4,11 +4,10 @@ import { Route, Link } from "react-router-dom";
 import axios from "axios";
 
 // Components
-import UserForm from "./UserForm.js";
+import UserFormTwo from "./UserFormTwo.js";
 import Dashboard from "./Dashboard.js";
 
 //material ui
-
 
 //rendering all protected components and keeping state here
 
@@ -123,7 +122,7 @@ class Secret extends Component {
         ) : (
           <div>
             <h4>Please Register To Access Secret</h4>
-            <UserForm postUserInfo={this.postUserInfo} />
+            <UserFormTwo postUserInfo={this.postUserInfo} />
           </div>
         )}
       </div>
